Added parallel analysis using Processes instead of Threads.
Now, both the frequencies and conservation estimations are done using
independent processes instead of threads. This way the GIL does not
limit the performance of this two stages.
There is still a lot to be done. Cleaning the algorithm while debugging
it will be the next tasks to be done.