Updated methods to calculate weights, frequencies and conservation.
Added new approach to take advantage of the multicore arquitecture
for calculate sequences' weights as well as position (or column)
frequencies and conservation.

At the moment there are some problems due to the way the
multiprocessing module works. Apparently the Pool.apply_async
function cannot call methods which are not at the top level of the
module.