Updated methods to calculate weights, frequencies and conservation.
In the end it was not necessary to declare functions at the top
level of the module to be called by the Pool.apply_async method,
but it does seem there are some problems when you try to call
statics methods. I will take a closer look at this in order to
fully understand it.

It is already clear by now, though, that using the multiprocessing
module instead of the multithreading one was the right thing to do
in order to improve the system's performance.