Finished updating the _Conservation_Index.py code.
It has been updated every algorithm which could be parallelized as
well as the ones from the Report class. In the Report class there
have been some modifications with the aim of improve the algorithms'
performance focusing on the way the strings are managed.

Because it is now considered a stable version of the system, the
reports have been regenerated using different settings to analyze
the set of sequences previously aligned.