Finished updating the _Conservation_Index.py code.
It has been updated every algorithm where we could work with
generators instead of iterators. This way we improve both
memory usage and execution time (it is needed a deeper analysis
about these stats).

Because it is now considered a stable version of the system, the
reports have been regenerated using different settings to analyze
the set of sequences previously aligned.